pyproject.toml
[tool.poetry]
name = "rl-robo-book-examples"
version = "0.1.0"
description = ""
authors = ["Shundo Kishi <syundo0730@gmail.com>"]

[[tool.poetry.source]]
name = "testpypi"
url = "https://test.pypi.org/simple/"

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.17"
machina-rl = "^0.2.1"
premaidai_gym = "0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
